Refactor CreateRegistration to its own namespace and folder
Moved CreateRegistration command, handler, validator, and related tests from the 'Create' namespace and folder to a new 'CreateRegistration' namespace and folder. Updated all relevant using statements and namespaces to reflect this change for improved clarity and organization.
